separate Datei für Wurzel
